requirements.txt
numpy==1.25.2
pandas==2.1.4
matplotlib==3.8.0
seaborn>=0.13
scipy==1.11.4
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
tqdm==4.66.1
pyyaml>=6.0                    
opencv-python>=4.8            
Pillow>=10.0
fastapi>=0.104                 
uvicorn>=0.24                
onnx>=1.15